requirements.txt
beautifulsoup4
requests
openai
faiss-cpu
fastapi
uvicorn
pydantic
numpy
